[IMP] estate, estate_account: Improved logic and function definition
Moved the invoice related functions to the estate_account model, redefined the logic for visit scheduler and removed the unnecessary visit end time field, and added a cron job to auto send reminders 1 day.

Chapter 13
