+++ estate_account/models/estate_properties.py
@@ -1,6 +1,6 @@
 import logging
 
-from odoo import fields, models
+from odoo import api, fields, models
 from odoo import Command
 
 _logger = logging.getLogger(__name__)
@@ -9,7 +9,7 @@ _logger = logging.getLogger(__name__)
 class EstateProperties(models.BaseModel):
     _inherit = 'estate.properties'
 
-    invoice_count = fields.Integer(compute='')
+    invoice_count = fields.Integer(compute='_compute_invoice_count')
 
 
     def _check_admin_fees(self, admin_fees):
@@ -55,3 +55,41 @@ class EstateProperties(models.BaseModel):
         # _logger.error(invoice_vals)
         self.env['account.move'].create(invoice_vals)
         return super()
+
+    def _find_invoices(self):
+        self.ensure_one()
+        check_name = "Property: " + self.display_name if self.display_name else None
+        if check_name and self.selling_price > 0:  # type: ignore
+            invoice_lines = self.env['account.move.line'].search([  # type: ignore
+                ('move_id.move_type', '=', 'out_invoice'),  # type: ignore
+                ('name', 'ilike', check_name),  # type: ignore
+            ])        
+            return invoice_lines.mapped('move_id').ids  # type: ignore
+        else:
+            return False
+
+    @api.depends('state')
+    def _compute_invoice_count(self):
+        # _logger.error("HELLOOOOOOOOOOO")
+        invoices = self._find_invoices()
+        for property in self:
+            if invoices:
+                property.invoice_count = len(invoices)
+            else:
+                property.invoice_count = 0
+
+    def action_view_partner_invoices(self):
+        if self.selling_price > 0:  # type: ignore
+            invoice_ids = self._find_invoices()
+    
+            if invoice_ids:
+                invoice_ids = int(invoice_ids[0])
+                action = self.env["ir.actions.actions"]._for_xml_id("account.action_move_out_invoice")  # type: ignore
+                action['res_id'] = invoice_ids
+                action['view_mode'] = 'form'
+                action['views'] = [(self.env.ref('account.view_move_form').id, 'form')]  # type: ignore
+                action['domain'] = [
+                    ('id', 'in', invoice_ids),
+                    ('partner_id', '=', self.buyer_id.id),  # type: ignore
+                ]
+                return action
